rev-py0206

+++ 0x02-python-import_modules/100-my_calculator.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
     elif operator == '*':
         answer = a * b
     elif operator == '/':
-        answer = a / b
+        answer = int(a / b)
     else:
         print("Unknown operator. Available operators: +, -, * and /")
         exit(1)
